salsify-lens/pyproject.toml
[tool.poetry]
name = "salsify-lens"
version = "0.1.0"
description = ""
authors = ["Xx <xx@xx>"]
readme = "README.md"
packages = [{include = "salsify_lens"}]

[tool.poetry.dependencies]
python = "3.11.9"
opencv-python = "^4.9.0.80"
numpy = "^1.24.3"
keras = "2.13.1"
tensorflow = "2.13.0"
pydantic = "1.10.13"
tensorflow-macos = "2.13.1"
pygame = "^2.5.2"
pillow = "^11.0.0"
scipy = "^1.14.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
